Move FluidSerializer code to runtime-utils
The only change from FluidSerializer -> FluidSerializerBase is
removing the 2 references to IFluidSerializer interface
